requirements.txt
# ============================================================================
# MAXIMUM REQUIREMENTS - Complete Trading Bot Dependencies
# ============================================================================
# This includes ALL dependencies for full functionality:
# - Core trading system
# - All broker/exchange integrations
# - Blockchain/DEX (Web3, Uniswap, Flash Loans)
# - AI/ML models
# - Web dashboard
# - Development tools
# - Testing framework
# - Security & profiling tools
# ============================================================================

# ============================================================================
# Core Trading Bot Dependencies
# ============================================================================

# High-performance data processing
polars>=1.0.0
pyarrow>=18.0.0
duckdb>=1.1.0

# Async runtime
uvloop>=0.21.0
asyncio>=3.4.3

# Fast serialization
msgspec>=0.18.0

# Structured logging
structlog>=24.0.0

# Observability
opentelemetry-api>=1.30.0
opentelemetry-sdk>=1.30.0
opentelemetry-exporter-otlp>=1.30.0

# Numerical computing
numpy>=2.0.0
scipy>=1.15.0

# Type extensions
typing-extensions>=4.12.0

# ============================================================================
# Exchange/Broker Connectivity (Essential for Arbitrage)
# ============================================================================

# Modern broker connectivity
alpaca-py>=0.34.0
httpx>=0.28.0  # Updated to 0.28+ for 2025 standards (no deprecated verify/cert patterns)
httpx-sse>=0.4.0
python-binance>=1.0.0
python-dotenv>=1.0.0
yfinance>=0.2.0
orjson>=3.9.0
websockets>=12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
keyring>=24.3.0

# Interactive Brokers
ibapi>=9.81.1.post1

# Additional exchanges
ccxt>=4.0.0
krakenex>=2.2.0

# ============================================================================
# Blockchain/DEX (Web3, Uniswap, Flash Loans)
# ============================================================================

# Web3 core
web3>=7.0.0
eth-abi>=5.0.0
eth-account>=0.13.0
eth-typing>=4.0.0
eth-utils>=5.0.0
hexbytes>=1.2.0

# GraphQL for subgraphs
gql[aiohttp]>=3.5.0
aiohttp>=3.9.0

# Contract interaction
eth-brownie>=1.20.0

# ============================================================================
# Web Dashboard (FastAPI/Uvicorn)
# ============================================================================

fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.0
python-multipart>=0.0.9
starlette>=0.40.0

# ============================================================================
# Machine Learning (AI Decider)
# ============================================================================

# Core ML
mlflow>=2.9.2
scikit-learn>=1.6.0
xgboost>=2.1.0
lightgbm>=4.0.0

# Deep learning (optional but included for maximum)
torch>=2.0.0
tensorflow>=2.15.0

# ML utilities
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
joblib>=1.4.0

# ============================================================================
# Data Analysis & Visualization
# ============================================================================

plotly>=5.18.0
dash>=2.14.0
jupyter>=1.0.0
jupyterlab>=4.0.0
notebook>=7.0.0
ipython>=8.20.0

# ============================================================================
# Database & Caching
# ============================================================================

redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
pymongo>=4.6.0

# ============================================================================
# Development Dependencies
# ============================================================================

# Linting and formatting
ruff>=0.9.0
black>=24.0.0
isort>=5.13.0
autopep8>=2.0.0

# Type checking
pyright>=1.1.400
mypy>=1.13.0
types-requests>=2.31.0
types-redis>=4.6.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.25.0
pytest-benchmark>=5.1.0
pytest-cov>=6.0.0
pytest-xdist>=3.6.0
pytest-mock>=3.12.0
pytest-timeout>=2.2.0
vcrpy>=6.0.0
hypothesis>=6.92.0
faker>=22.0.0

# Security
bandit[toml]>=1.8.0
pip-audit>=2.8.0
safety>=3.0.0

# Profiling & Performance
py-spy>=0.4.0
scalene>=1.5.0
memory-profiler>=0.61.0
line-profiler>=4.1.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=2.0.0
mkdocs>=1.5.0
mkdocs-material>=9.5.0

# ============================================================================
# Utility Libraries
# ============================================================================

# CLI tools
click>=8.1.0
rich>=13.7.0
typer>=0.12.0

# Date/Time
python-dateutil>=2.8.0
pytz>=2024.1

# Environment & Config
environs>=11.0.0
dynaconf>=3.2.0

# HTTP clients
requests>=2.31.0
aiofiles>=23.2.0

# Cryptography
cryptography>=42.0.0
pycryptodome>=3.20.0

# Task queue
celery>=5.3.0
dramatiq>=1.16.0

# Monitoring
prometheus-client>=0.20.0
sentry-sdk>=2.0.0

# ============================================================================
# Native Extensions (for performance)
# ============================================================================

maturin>=1.7.0
nanobind>=2.4.0
cython>=3.0.0

# ============================================================================
# Additional Utilities
# ============================================================================

# Data validation
marshmallow>=3.20.0
cerberus>=1.3.5

# Rate limiting
ratelimit>=2.2.1

# Retry logic
tenacity>=8.2.0
backoff>=2.2.1

# UUID generation
shortuuid>=1.0.11

# File operations
pathlib2>=2.3.7
watchdog>=4.0.0

# Compression
zstandard>=0.22.0

# Parsing
python-rapidjson>=1.14
simplejson>=3.19.0
